requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
Faker==35.2.0
fastapi==0.115.8
ffmpeg-python==0.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
markdown2==2.5.3
numpy>1.23.0
packaging==24.2
pillow==11.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
srt==3.5.3
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
vosk==0.3.45
websockets==14.2
# Logging utilities
loguru==0.7.2  # Advanced logging
